App working with save itinerary functionality to TripScreen.kt complete, now trying to add save itineraries to firestore and navigate from TripScreen.kt to ItineraryScreen.kt when clicking any saved itinerary card.

+++ app/src/main/java/tfg/sal/tripl/appcontent/home/data/network/poi/OpenTripMapClient.kt
@@ -10,7 +10,7 @@ interface OpenTripMapClient {
     suspend fun getPOIList(
         @Query("format") format: String,
         @Query("lang") lang: String,
-        @Query("radius") radius: Double,
+        @Query("radius") radius: Float,
         @Query("lat") lat: Double,
         @Query("lon") lon: Double,
         @Query("apikey") apikey: String
