+++ app/src/main/java/tfg/sal/tripl/appcontent/home/data/poi/OpenTripMapRepository.kt
@@ -4,7 +4,7 @@ import tfg.sal.tripl.appcontent.home.data.network.poi.OpenTripMapService
 import javax.inject.Inject
 
 class OpenTripMapRepository @Inject constructor(private val api: OpenTripMapService) {
-    suspend fun getPOIList(radius: Double, lat: Double, lon: Double): PointsOfInterest {
+    suspend fun getPOIList(radius: Float, lat: Double, lon: Double): PointsOfInterest {
         return api.getPOIList(radius, lat, lon)
     }
 }
